pyproject.toml
[project]
name = "hackaton-mistral-episen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google>=3.0.0",
    "google-analytics-data>=0.18.19",
    "mcp>=1.14.0",
    "pydantic",
    "requests",
    "python-dotenv",
    "mistralai"
]
